+++ CityStats/Systems/ModUISystem.cs
@@ -12,13 +12,33 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 
 namespace CityStats.Systems {
-    internal partial class ModUISystem : UISystemBase {
+    internal partial class ModUISystem : UISystemBase, IDefaultSerializable, ISerializable {
+        private const string defaultHiddenStats = "";
+        private const int SAVE_VERSION = 0;
+
         private ProxyAction togglePanelBindingAction;
 
+        // TODO: Once multiple save formats exist, consider tracking all formats in constants
+        // Source: https://github.com/krzychu124/Traffic/blob/main/Code/Systems/DataMigration/TrafficDataMigrationSystem.cs
+
+        /// <summary>
+        /// Loaded save version format
+        /// </summary>
+        private int loadedSaveVersion = 0;
+        /// <summary>
+        /// Current/target save version format
+        /// </summary>
+        private const int targetSaveVersion = 0;
+
+        /// <summary>
+        /// Hidden stat icons (as comma-separated string)
+        /// </summary>
+        /// <remarks>
+        /// Serialized and deserialized automatically within game save
+        /// </remarks>
         private ValueBinding<string> hiddenStatsBinding;
         private ValueBinding<Vector2> panelPositionBinding;
         private ValueBinding<bool> panelVisibleBinding;
-        // TODO: See if this can be changed to a 'GetterValueBinding' (previously crashed with writing failure!)
         private ValueBinding<StatsPanelOrientation> panelOrientationBinding;
 
 
@@ -35,7 +55,7 @@ namespace CityStats.Systems {
             // Action Phases: Performed, Started, Cancelled, Waiting, Disabled
 
             // Value bindings
-            hiddenStatsBinding = new ValueBinding<string>(Mod.NAME, UIBindingData.VALUE_HIDDEN_STATS, "");
+            hiddenStatsBinding = new ValueBinding<string>(Mod.NAME, UIBindingData.VALUE_HIDDEN_STATS, defaultHiddenStats);
             AddBinding(hiddenStatsBinding);
             panelVisibleBinding = new ValueBinding<bool>(Mod.NAME, UIBindingData.VALUE_PANEL_VISIBLE, false);
             AddBinding(panelVisibleBinding);
@@ -57,7 +77,7 @@ namespace CityStats.Systems {
             AddBinding(setHiddenStatsTrigger);
             var setPanelVisibleTrigger = new TriggerBinding<bool>(Mod.NAME, UIBindingData.TRIGGER_SET_PANEL_VISIBLE, SetPanelVisibility);
             AddBinding(setPanelVisibleTrigger);
-            var setPanelPositionTrigger = new TriggerBinding<Vector2>(Mod.NAME, UIBindingData.TRIGGER_SET_PANEL_POSITION, OnSetPanelPositionTrigger);
+            var setPanelPositionTrigger = new TriggerBinding<Vector2>(Mod.NAME, UIBindingData.TRIGGER_SET_PANEL_POSITION, SetPanelPosition);
             AddBinding(setPanelPositionTrigger);
 
             Mod.Settings.onSettingsApplied += OnModSettingsApplied;
@@ -87,6 +107,7 @@ namespace CityStats.Systems {
         #endregion
 
 
+        #region Methods
         /// <summary>
         /// Handle mod setting changes (applied immediately upon change, not when closing menu)
         /// </summary>
@@ -105,7 +126,10 @@ namespace CityStats.Systems {
         }
 
 
-        private void OnSetPanelPositionTrigger(Vector2 position) {
+        /// <summary>
+        /// Update stats panel position
+        /// </summary>
+        private void SetPanelPosition(Vector2 position) {
             panelPositionBinding.Update(position);
         }
 
@@ -140,7 +164,7 @@ namespace CityStats.Systems {
         /// Uses a comma-separated string to simplify future serialization within save game (via ECS Entity)
         /// </remarks>
         public void SetHiddenStats(string stats) {
-            Mod.Log.Debug($"Updating hidden stats ({stats})");
+            Mod.Log.Debug($"[{nameof(ModUISystem)}] Updating hidden stats ({stats})");
             hiddenStatsBinding.Update(stats);
         }
 
@@ -154,11 +178,78 @@ namespace CityStats.Systems {
 
 
         /// <summary>
+        /// Reset hidden stats (to display all)
+        /// </summary>
+        public void ResetHiddenStats() {
+            hiddenStatsBinding.Update(defaultHiddenStats);
+            Mod.Log.Info($"[{nameof(ModUISystem)}] Reset hidden stats");
+        }
+
+
+        /// <summary>
         /// Reset stats panel position (if inaccessible, etc)
         /// </summary>
         public void ResetPanelPosition() {
             panelPositionBinding.Update(Vector2.zero);
             Mod.Log.Info($"[{nameof(ModUISystem)}] Reset panel position");
         }
+        #endregion
+
+
+        #region Serialization
+        /// <summary>
+        /// Set default values (only when not previously saved)
+        /// </summary>
+        public void SetDefaults(Context context) {
+            Mod.Log.Debug($"[{nameof(ModUISystem)}] SetDefaults");
+
+            loadedSaveVersion = 0;
+            hiddenStatsBinding.Update(defaultHiddenStats);
+        }
+
+        /// <summary>
+        /// Store system data within save game
+        /// </summary>
+        /// <remarks>
+        /// NOTE: Save order must remain consistent for compatibility, and be mirrored when reading! If save data
+        ///         changes in a non-compatible way (ie. reordering/replacing existing value types), the save
+        ///         version must be updated and the migration handled!
+        /// </remarks>
+        public void Serialize<TWriter>(TWriter writer) where TWriter : IWriter {
+            // CAUTION: Ensure that any changes to serialization order/type are handled appropriately, including adding
+            //            a save version format upgrade if necessary!
+
+            Mod.Log.Debug($"[{nameof(ModUISystem)}] Serializing (v{targetSaveVersion}) (hiddenStats={hiddenStatsBinding.value})");
+
+            writer.Write(targetSaveVersion);
+
+            writer.Write(hiddenStatsBinding.value);
+
+            Mod.Log.Debug($"[{nameof(ModUISystem)}] Serialized");
+        }
+
+        /// <summary>
+        /// Read system data from save game
+        /// </summary>
+        /// <remarks>
+        /// NOTE: Save order must remain consistent for compatibility, and be mirrored when storing! If save data
+        ///         changes in a non-compatible way (ie. reordering/replacing existing value types), the save
+        ///         version must be updated and the migration handled!
+        /// </remarks>
+        public void Deserialize<TReader>(TReader reader) where TReader : IReader {
+            // CAUTION: Ensure that any changes to deserialization order/type are handled appropriately, including adding
+            //            a save version format upgrade if necessary!
+
+            Mod.Log.Debug($"[{nameof(ModUISystem)}] Deserializing");
+
+            reader.Read(out loadedSaveVersion);
+
+            string savedHiddenStats;
+            reader.Read(out savedHiddenStats);
+            hiddenStatsBinding.Update(savedHiddenStats);
+
+            Mod.Log.Debug($"[{nameof(ModUISystem)}] Deserialized (v{loadedSaveVersion}) (hiddenStats={savedHiddenStats})");
+        }
+        #endregion
     }
 }